Move I18nComponent to correct package
